Add class to handle list of sample characters
This commit introduces SampleList as an object to store a list of random
characters.

The main idea is that it doesn't need to store all data's characters,
only a fraction of it would be valid.

SampleList has a capacity defined and the list items can be replaced
at maximum 100 times.